Add properties for version label as number and also "next by distance for each version component"
The intent of these changes is to open up an effective build number (via version label and "next by distance" based on the number of commits since last tag matching the version pattern). In our model we wish to use the label for this (hence the addition of version label as number) but it is entirely reasonable to treat the patch number as auto-incrementing with each commit the same way.